+++ src/main/java/lotto/controller/LottoController.java
@@ -3,6 +3,7 @@ package lotto.controller;
 import lotto.dto.LottoDto;
 import lotto.domain.*;
 import lotto.parser.InputParser;
+import lotto.service.LottoRankService;
 import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -15,11 +16,18 @@ public class LottoController {
     private final InputView inputView;
     private final OutputView outputView;
     private final LottoService lottoService;
+    private final LottoRankService lottoRankService;
 
-    public LottoController(InputView inputView, OutputView outputView, LottoService lottoService) {
+    public LottoController(
+            InputView inputView,
+            OutputView outputView,
+            LottoService lottoService,
+            LottoRankService lottoRankService
+    ) {
         this.inputView = inputView;
         this.outputView = outputView;
         this.lottoService = lottoService;
+        this.lottoRankService = lottoRankService;
     }
 
     public void run() {
@@ -35,14 +43,14 @@ public class LottoController {
         BonusNumber bonusNumber = readBonusNumber();
         checkDuplicate(winningNumbers, bonusNumber);
 
-        lottos.forEach(lotto -> lottoService.determineRank(lotto, winningNumbers, bonusNumber));
+        lottos.forEach(lotto -> lottoRankService.determineRank(lotto, winningNumbers, bonusNumber));
 
         int totalPurchase = purchaseAmount.getPurchaseAmount();
-        double profitRate = lottoService.calculateProfitRate(totalPurchase);
+        double profitRate = lottoRankService.calculateProfitRate(totalPurchase);
 
         outputView.showWinningStatisticsTitle();
-        Map<LottoRank, Integer> ranks = lottoService.getRanks();
-        outputView.showWinningStatistics(ranks);
+        Map<LottoRank, Integer> lottoRanks = lottoRankService.getRanks();
+        outputView.showWinningStatistics(lottoRanks);
         outputView.showProfitRate(profitRate);
 
         inputView.close();
