+++ src/main/java/lotto/service/LottoService.java
@@ -5,7 +5,6 @@ import lotto.domain.*;
 import lotto.exception.ErrorCode;
 
 import java.util.List;
-import java.util.Map;
 import java.util.stream.IntStream;
 
 public class LottoService {
@@ -14,13 +13,6 @@ public class LottoService {
     private static final int LOTTO_NUMBER_MIN = 1;
     private static final int LOTTO_NUMBER_MAX = 45;
     private static final int LOTTO_NUMBER_COUNT = 6;
-    private static final long INITIAL_TOTAL_PRIZE = 0L;
-
-    private final LottoRankCounter lottoRankCounter;
-
-    public LottoService(LottoRankCounter lottoRankCounter) {
-        this.lottoRankCounter = lottoRankCounter;
-    }
 
     public List<Lotto> purchase(PurchaseAmount purchaseAmount) {
         int purchaseCount = purchaseAmount.getPurchaseAmount() / LOTTO_PRIZE;
@@ -40,43 +32,8 @@ public class LottoService {
         }
     }
 
-    public void determineRank(Lotto lotto, WinningNumbers winningNumbers, BonusNumber bonusNumber) {
-        int matchingCount = countMatchingNumbers(lotto, winningNumbers);
-        boolean isBonusNumberMatched = isBonusNumberMatched(lotto, bonusNumber);
-
-        lottoRankCounter.update(matchingCount, isBonusNumberMatched);
-    }
-
-    public Map<LottoRank, Integer> getRanks() {
-        return lottoRankCounter.getLottoResult();
-    }
-
-    public double calculateProfitRate(int totalPurchase) {
-        Map<LottoRank, Integer> lottoResult = lottoRankCounter.getLottoResult();
-
-        Long totalPrize = lottoResult.entrySet()
-                .stream()
-                .filter(entry -> entry.getValue() != 0)
-                .map(entry -> entry.getKey().getPrize() * entry.getValue())
-                .reduce(INITIAL_TOTAL_PRIZE, Long::sum);
-
-        return (double) totalPrize / totalPurchase * 100;
-    }
-
-    public long payPrize(LottoRank lottoRank) {
-        return lottoRank.getPrize();
-    }
-
     private Lotto generateLotto(int start, int end, int count) {
         List<Integer> numbers = Randoms.pickUniqueNumbersInRange(start, end, count);
         return new Lotto(numbers);
-    }
-
-    private int countMatchingNumbers(Lotto lotto, WinningNumbers winningNumbers) {
-        return lotto.countMatchingNumbers(winningNumbers);
-    }
-
-    private boolean isBonusNumberMatched(Lotto lotto, BonusNumber bonusNumber) {
-        return lotto.isBonusNumberMatched(bonusNumber);
     }
 }
